requirements.txt
# Core Django
Django>=5.0,<6.0
djangorestframework>=3.14.0
django-cors-headers>=4.3.0
django-filter>=23.3
drf-spectacular>=0.26.0
djangorestframework-simplejwt>=5.3.0
python-decouple>=3.8
whitenoise>=6.6.0
gunicorn>=21.2.0

# Database
psycopg2-binary>=2.9.9

# Utilities
Pillow>=10.1.0
openpyxl>=3.1.2
reportlab>=4.0.0

# Testing
hypothesis>=6.0.0
pytest-django>=4.5.0
